+++ src/Common/Services/OrderItemService.js
@@ -1,14 +1,14 @@
 import Parse from "parse";
 
 // CREATE operation - new orderItem with Name
-export const createOrderItem = (foodPointer, user) => {
-    console.log("foodPointer: ", foodPointer);
-    console.log("user: ", user);
+export const createOrderItem = (foodName, userEmail) => {
+    console.log("foodName: ", foodName);
+    console.log("userEmail: ", userEmail);
     const OrderItem = Parse.Object.extend("OrderItem");
     const orderItem = new OrderItem();
-    // using setter to UPDATE the object food pointer and user email
-    orderItem.set("food", foodPointer);
-    orderItem.set("user", user);
+    // using setter to UPDATE the object food name and user email
+    orderItem.set("food", foodName);
+    orderItem.set("user", userEmail);
 
     return orderItem.save().then((result) => {
       // returns new OrderItem object
@@ -16,11 +16,11 @@ export const createOrderItem = (foodPointer, user) => {
     });
   };
 
-// READ operation - get all orderItems with userEmail in Parse class OrderItem
-export const getAllOrderItemsByUser = (user) => {
+// READ operation - get all orderItems with user id in Parse class OrderItem
+export const getAllOrderItemsByUser = (userEmail) => {
     const OrderItem = Parse.Object.extend("OrderItem");
     const query = new Parse.Query(OrderItem);
-    query.equalTo("user", user);
+    query.equalTo("user", userEmail);
     return query.find().then((results) => {
       // returns array of OrderItem objects
       console.log("results: ", results);
@@ -28,11 +28,22 @@ export const getAllOrderItemsByUser = (user) => {
     });
   };
 
-// DELETE operation - remove orderItems by userEmail
-export const removeAllOrderItemsByUser = (user) => {
+// READ operation - get all Object Parse class OrderItem objects
+export const getAllOrderItems = () => {
     const OrderItem = Parse.Object.extend("OrderItem");
     const query = new Parse.Query(OrderItem);
-    query.equalTo("user", user);
+    return query.find().then((results) => {
+        // returns array of OrderItem objects
+        return results;
+      });
+  };
+
+
+// DELETE operation - remove orderItems by userEmail
+export const removeAllOrderItemsByUser = (userEmail) => {
+    const OrderItem = Parse.Object.extend("OrderItem");
+    const query = new Parse.Query(OrderItem);
+    query.equalTo("user", userEmail);
 
     return query.find().then((orderItems) => {
       orderItems.destroyAll();